modules/location_ingest/requirements.txt
protobuf
grpcio
kafka-python